Use gsize for internal functions. buffer length is a gint, not a guint. We
	* src/pack.c: Use gsize for internal functions.
	    buffer length is a gint, not a guint.  We return
	    the number of bytes packed/unpacked, which can be
	    -1.  So the buffer can't be more then a gint.
	* src/udp.c (gnet_udp_socket_receive/send): "
	* src/mcast.c (gnet_mcast_socket_receive/send): "
	* src/iochannel.[ch]: Use gsize instead of guint

+++ src/iochannel.h
@@ -29,22 +29,22 @@ extern "C" {
 
 GIOError gnet_io_channel_writen (GIOChannel*   channel, 
 				 gpointer      buffer, 
-				 guint         length,
-				 guint*        bytes_writtenp);
+				 gsize         length,
+				 gsize*        bytes_writtenp);
 
 GIOError gnet_io_channel_readn (GIOChannel*    channel, 
 				 gpointer      buffer, 
-				 guint         length,
-				 guint*        bytes_readp);
+				 gsize         length,
+				 gsize*        bytes_readp);
 
 GIOError gnet_io_channel_readline (GIOChannel* channel, 
 				   gchar*      buffer, 
-				   guint       length,
-				   guint*      bytes_readp);
+				   gsize       length,
+				   gsize*      bytes_readp);
 
 GIOError gnet_io_channel_readline_strdup (GIOChannel*   channel, 
 					  gchar**       bufferp, 
-					  guint*        bytes_readp);
+					  gsize*        bytes_readp);
 
 
 #ifdef __cplusplus
